Update script.js
Added security feature:
A user cannot activate any keys for activating inspect element to avoid inspecting and editing fake results and such. (0% rate in terms of scamming).

Update:
The functions or scripts of the login form index and the user-dashboard is merged so that it is very easy to understand, read and maintain.